rename kungfuconfig to kungfu/config
