+++ srcs/go/libkungfu-comm/main.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/lsds/KungFu/srcs/go/kungfu/peer"
 	kb "github.com/lsds/KungFu/srcs/go/kungfu/base"
-	kc "github.com/lsds/KungFu/srcs/go/kungfuconfig"
+	kc "github.com/lsds/KungFu/srcs/go/kungfu/config"
 	"github.com/lsds/KungFu/srcs/go/log"
 	"github.com/lsds/KungFu/srcs/go/utils"
 )
