+++ srcs/go/monitor/counters_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"testing"
 
-	kc "github.com/lsds/KungFu/srcs/go/kungfuconfig"
+	kc "github.com/lsds/KungFu/srcs/go/kungfu/config"
 	"github.com/lsds/KungFu/srcs/go/plan"
 )
