+++ srcs/go/monitor/monitor.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 	"time"
 
-	kc "github.com/lsds/KungFu/srcs/go/kungfuconfig"
+	kc "github.com/lsds/KungFu/srcs/go/kungfu/config"
 	"github.com/lsds/KungFu/srcs/go/plan"
 )
