+++ srcs/go/rchannel/connection/connection.go
@@ -7,7 +7,7 @@ import (
 	"sync"
 	"time"
 
-	kc "github.com/lsds/KungFu/srcs/go/kungfuconfig"
+	kc "github.com/lsds/KungFu/srcs/go/kungfu/config"
 	"github.com/lsds/KungFu/srcs/go/log"
 	"github.com/lsds/KungFu/srcs/go/plan"
 )
