+++ srcs/go/rchannel/server/server.go
@@ -8,7 +8,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	kc "github.com/lsds/KungFu/srcs/go/kungfuconfig"
+	kc "github.com/lsds/KungFu/srcs/go/kungfu/config"
 	"github.com/lsds/KungFu/srcs/go/log"
 	"github.com/lsds/KungFu/srcs/go/plan"
 	"github.com/lsds/KungFu/srcs/go/rchannel/connection"
